+++ pkg/middleware/ctxtags/ctxtags.go
@@ -1,4 +1,4 @@
-package grpc_ctxtags
+package ctxtags
 
 import (
 	"context"
@@ -41,7 +41,7 @@ type WrappedServerStream struct {
 	WrappedContext context.Context
 }
 
-// Context returns the wrapper's WrappedContext, overwriting the nested grpc.ServerStream.Context()
+// Context returns the wrapper's WrappedContext, overwriting the nested grpc.ServerStream.Context().
 func (w *WrappedServerStream) Context() context.Context {
 	return w.WrappedContext
 }
